+++ src/main/java/com/bison/petsyApp/dtos/utilisateur/FindUtilisateurDTO.java
@@ -11,7 +11,7 @@ import java.util.Collection;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FindUtilisateurDTO {
-    private int id_utilisateur;
+    private Long id_utilisateur;
     private String nom_utilisateur;
     private String prenom_utilisateur;
     private String email_utilisateur;
