+++ src/main/java/com/bison/petsyApp/models/Utilisateur.java
@@ -15,7 +15,7 @@ import java.util.Collection;
 public class Utilisateur {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_utilisateur;
+    private Long id_utilisateur;
     private String nom_utilisateur;
     private String prenom_utilisateur;
     private String email_utilisateur;
