+++ src/main/java/com/bison/petsyApp/repositories/UtilisateurRepository.java
@@ -3,5 +3,5 @@ package com.bison.petsyApp.repositories;
 import com.bison.petsyApp.models.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UtilisateurRepository extends JpaRepository<Utilisateur, Integer> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur, Long> {
 }
